# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TrossApp API",
    "version": "1.0.0",
    "description": "\n        TrossApp Backend REST API - Skills-based work order management system.\n        \n        This documentation covers foundational authentication, user management, \n        and role-based access control endpoints.\n        \n        ## Authentication\n        \n        Most endpoints require a Bearer token in the Authorization header:\n        ```\n        Authorization: Bearer <your-jwt-token>\n        ```\n        \n        ## Dual Authentication Strategy\n        \n        - **Development Mode:** JWT-based authentication with test users\n        - **Production Mode:** Auth0 OAuth2/OIDC integration\n        \n        ## Response Format\n        \n        All successful responses follow this format:\n        ```json\n        {\n          \"success\": true,\n          \"data\": {...},\n          \"timestamp\": \"2025-11-12T12:00:00.000Z\"\n        }\n        ```\n        \n        Error responses:\n        ```json\n        {\n          \"success\": false,\n          \"error\": \"Error message describing what went wrong\",\n          \"timestamp\": \"2025-11-12T12:00:00.000Z\"\n        }\n        ```\n      ",
    "contact": {
      "name": "TrossApp Team",
      "email": "dev@trossapp.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
...
    },
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from /api/dev/token (dev) or Auth0 (prod)"
      }
    },
...
  },
...
  "paths": {
...
    "/api/technicians": {
      "get": {
        "tags": [
          "Technicians"
        ],
        "summary": "Get all technicians",
        "description": "Retrieve a paginated list of technicians. Customers see limited info (field-level filtering applies).",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 50
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "default": "created_at"
            }
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "DESC"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Technicians retrieved successfully"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "Technicians"
        ],
        "summary": "Create new technician (manager+ only)",
        "description": "Create a technician profile. Typically done during hiring process.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "license_number"
                ],
                "properties": {
                  "license_number": {
                    "type": "string"
                  },
                  "hourly_rate": {
                    "type": "number"
                  },
                  "status": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Technician created successfully"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden - Manager+ access required"
          }
        }
      }
    },
    "/api/technicians/{id}": {
      "get": {
        "tags": [
          "Technicians"
        ],
        "summary": "Get technician by ID",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Technician retrieved successfully"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Technician not found"
          }
        }
      },
      "patch": {
        "tags": [
          "Technicians"
        ],
        "summary": "Update technician",
        "description": "Technicians can update their own profile, manager+ can update any technician.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Technician updated successfully"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Technician not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Technicians"
        ],
        "summary": "Deactivate technician (manager+ only)",
        "description": "Soft delete (deactivate) a technician account.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Technician deactivated successfully"
          },
          "403": {
            "description": "Forbidden - Manager+ access required"
          },
          "404": {
            "description": "Technician not found"
          }
        }
      }
    },
...
  }
}